move pin defs to own header

+++ digipot.h
@@ -1,4 +1,3 @@
-#include <Arduino.h>
 #ifndef DIGIPOT_H
 #define DIGIPOT_H
 
@@ -6,6 +5,6 @@ void digipotWrite(int address, int value);
 
 void digipotInit();
 
-void digipotSetVoltage(uint8_t driver, int current);
+void digipotSetVoltage(int driver, int current);
 
 #endif
